Remove metadata.ts file
