+++ frontend/src/app/layout.tsx
@@ -1,9 +1,18 @@
-"use client";
-
 import "./globals.css";
 import Header from "@/components/Header";
 import StatusChecker from "@/components/StatusChecker";
 import { ToastProvider } from "@/contexts/ToastContext";
+import type { Metadata } from "next";
+
+const isGithubPages = process.env.NODE_ENV === "production" && process.env.GITHUB_PAGES === "true";
+
+export const metadata: Metadata = {
+  title: "FluentAI",
+  description: "Learning languages in a flash.",
+  icons: {
+    icon: `${isGithubPages ? '/FluentAI' : ''}/logo.png`,
+  }
+};
 
 export default function RootLayout({
   children,
