+++ frontend/src/contexts/ToastContext.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { createContext, useContext, useState, useCallback } from 'react';
 import { Toast, ToastContainer, ToastType } from '@/components/ui/Toast';
